refactor: separa actions
